Drop the concept of using URLs for all storages
This idea was good on paper, but not as great in practice.

Storages will often requires other parameters as input (e.g.: TLS
configuration for HTTPS-based storages). That aside, representing file
paths as URLs is awkward when a simple `PathBuf` is the more obvious
choice.

This uses moves all storage definition into the previous `Metadata`
type, but renames it into `Definition` to avoid confusion with actual
collection metadata.