Drop separate context_path field
Once we resolve the right context path for a caldav server, that becomes
our base_url, and the URL originally provided as input becomes
irrelevant.

Note that this simplification has unveiled a bug too; when the context
path is resolved via DNS, we're ignoring the domain/port found via SRV
records.