examples: use single thread runtime
The multithread runtime is a separate feature that is disabled by
default. This actually worked by coincidence. Or rather, a bug in Cargo?
Or Rust? Somewhere. It seems that because other crates in the same
repository rely on the multithreaded runtime feature, it gets compiled
if all crates are compiled together and things work. Odd.