Merge Collection trait into Storage trait
Move all the implementation-specific methods from Collection into
Storage and make Collection a struct instead.

The main reason for the previous approach was so that if StorageA
returned a Collection, one could not accidentally try to use this
collection with StorageB. However, such usage would be a semantic
programming error, not a something that a user can mess up, and trying
to protect from it was degrading the quality (both readability and
performance) of the entire library.

Collections mostly ended up having an `Arc` to their storage or cloning
it entirely. Additionally, Collections had to always be `Boxed`, which
resulted in all the Storage code having to wrap their results (even in
contexts where this was not necessary). Any code related to them being a
horrible mess and had an extra layer of indirection.

Collections are now a thin wrapper around an `href`. All storage
implementation logic is now in the Storage trait.