libdav: parametrise end event
So far each parser ended its work when it reached an `end` node for a
given tag. This generally corresponded to the start tag that the
previous parser had matched before delegating to the current parser, but
resulted in a brittle interface, where both the caller and callee were
required to have matching tags.

Make the end event for parsers parametrisable, so that the start and end
events are both defined in the same scope. This improves compostability
of parser and will help clean up some rather ugly bits.